+++ src/components/analytics/controller.ts
@@ -1,14 +1,27 @@
 import { CREATED, OK } from "http-status";
+import { Analytics } from "@prisma/client";
+import { AnalyticsServices } from "./services";
 import { apiResponse } from "@/helpers/apiResponse";
 import { AnalyticsDAO } from "@/db/analytics";
 import { AnalyticsCreateRequest } from "@/types/request/analytics";
+import { extractResourceSearchQueryParams } from "@/helpers/queryParams";
 
 export class AnalyticsController {
-  static async index(_req: Req, res: Res, next: NextFn) {
+  static async index(req: Req, res: Res, next: NextFn) {
     try {
-      const analytics = await AnalyticsDAO.findMany();
+      const { page, itemsPerPage, search, orderBy } =
+        extractResourceSearchQueryParams<Analytics>(req);
 
-      res.status(OK).send(apiResponse(analytics));
+      const analyticsServices = new AnalyticsServices();
+      const result = await analyticsServices.getAll(
+        page,
+        itemsPerPage,
+        search,
+        orderBy,
+      );
+      const total = await AnalyticsDAO.count;
+
+      res.status(OK).send(apiResponse({ result, total }));
     } catch (e) {
       next(e);
     }
@@ -16,7 +29,7 @@ export class AnalyticsController {
 
   static async show(req: Req, res: Res, next: NextFn) {
     try {
-      const analytics = await AnalyticsDAO.getById(req.params.id);
+      const analytics = await AnalyticsDAO._getById(req.params.id);
 
       res.status(OK).send(apiResponse(analytics));
     } catch (e) {
@@ -30,6 +43,17 @@ export class AnalyticsController {
       const analytics = await AnalyticsDAO.create(data);
 
       res.status(CREATED).send(apiResponse(analytics));
+    } catch (e) {
+      next(e);
+    }
+  }
+
+  static async summary(_req: Req, res: Res, next: NextFn) {
+    try {
+      const analyticsServices = new AnalyticsServices();
+      const summary = await analyticsServices.summary();
+
+      res.status(OK).send(apiResponse(summary));
     } catch (e) {
       next(e);
     }