+++ src/config/index.ts
@@ -111,8 +111,9 @@ const CONFIG = {
     },
     ENVIRONMENTS: {
       TEST: "test",
-      DEV: "dev",
+      DEV: "development",
       PRODUCTION: "production",
+      STAGING: "staging",
     },
     PLAYER_WELCOME_MESSAGE: process.env.PLAYER_WELCOME_MESSAGE || "",
   },
@@ -139,4 +140,5 @@ export enum PLAYER_STATUS {
   ACTIVE = "ACTIVO",
   BANNED = "BLOQUEADO",
 }
+
 export default CONFIG;